+++ packages/etherplay-alchemy/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @etherplay/alchemy
 
+## 0.0.14
+
+### Patch Changes
+
+- support different origin
+
 ## 0.0.13
 
 ### Patch Changes